Split and barrel cloud module for merge.
