+++ atst/jobs.py
@@ -4,14 +4,15 @@ import pendulum
 from atst.database import db
 from atst.queue import celery
 from atst.models import EnvironmentRole, JobFailure
-from atst.domain.csp.cloud import CloudProviderInterface, GeneralCSPException
+from atst.domain.csp.cloud.exceptions import GeneralCSPException
+from atst.domain.csp.cloud import CloudProviderInterface
 from atst.domain.applications import Applications
 from atst.domain.environments import Environments
 from atst.domain.portfolios import Portfolios
 from atst.domain.environment_roles import EnvironmentRoles
 from atst.models.utils import claim_for_update
 from atst.utils.localization import translate
-from atst.domain.csp.cloud import ApplicationCSPPayload
+from atst.domain.csp.cloud.models import ApplicationCSPPayload
 
 
 class RecordFailure(celery.Task):
